requirements.txt
# Core Streamlit and Web Framework
streamlit

# LangChain and AI Components
langchain
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langsmith

# Memory and Storage
langmem

# Data Validation and Configuration
pydantic
pyyaml

# Testing Framework
pytest
pytest-anyio

# Type Checking
mypy

# Visualization and Diagrams
pygraphviz

# Additional Utilities
numpy  # Used in advanced evaluation
hashlib  # Built-in, but good to document